requirements.txt
fastapi
uvicorn
playwright
python-dotenv
openai
httpx
